# Switch to dbtest.NewDB in batches codebase