--- a/enterprise/internal/batches/background/main_test.go
+++ b/enterprise/internal/batches/background/main_test.go
@@ -1,7 +1,0 @@
-package background
-
-import "github.com/sourcegraph/sourcegraph/internal/database/dbtesting"
-
-func init() {
-	dbtesting.DBNameSuffix = "batchesbackgrounddb"
-}