--- a/enterprise/internal/batches/resolvers/code_host_connection_test.go
+++ b/enterprise/internal/batches/resolvers/code_host_connection_test.go
@@ -16,7 +16,7 @@ import (
 	btypes "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/types"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/database"
-	"github.com/sourcegraph/sourcegraph/internal/database/dbtesting"
+	"github.com/sourcegraph/sourcegraph/internal/database/dbtest"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/auth"
 )
@@ -27,7 +27,7 @@ func TestCodeHostConnectionResolver(t *testing.T) {
 	}
 
 	ctx := backend.WithAuthzBypass(context.Background())
-	db := dbtesting.GetDB(t)
+	db := dbtest.NewDB(t, "")
 
 	pruneUserCredentials(t, db, nil)
 
@@ -154,7 +154,7 @@ func TestCodeHostConnectionResolver(t *testing.T) {
 	})
 
 	t.Run("User.BatchChangesCodeHosts", func(t *testing.T) {
-		cred, err := cstore.UserCredentials().Create(ctx, database.UserCredentialScope{
+		userCred, err := cstore.UserCredentials().Create(ctx, database.UserCredentialScope{
 			Domain:              database.UserCredentialDomainBatches,
 			ExternalServiceID:   ghRepo.ExternalRepo.ServiceID,
 			ExternalServiceType: ghRepo.ExternalRepo.ServiceType,
@@ -188,10 +188,10 @@ func TestCodeHostConnectionResolver(t *testing.T) {
 				ExternalServiceURL:  ghRepo.ExternalRepo.ServiceID,
 				ExternalServiceKind: extsvc.TypeToKind(ghRepo.ExternalRepo.ServiceType),
 				Credential: apitest.BatchChangesCredential{
-					ID:                  string(marshalBatchChangesCredentialID(cred.ID, false)),
-					ExternalServiceKind: extsvc.TypeToKind(cred.ExternalServiceType),
-					ExternalServiceURL:  cred.ExternalServiceID,
-					CreatedAt:           cred.CreatedAt.Format(time.RFC3339),
+					ID:                  string(marshalBatchChangesCredentialID(userCred.ID, false)),
+					ExternalServiceKind: extsvc.TypeToKind(userCred.ExternalServiceType),
+					ExternalServiceURL:  userCred.ExternalServiceID,
+					CreatedAt:           userCred.CreatedAt.Format(time.RFC3339),
 					IsSiteCredential:    false,
 				},
 			},