--- a/enterprise/internal/batches/service/main_test.go
+++ b/enterprise/internal/batches/service/main_test.go
@@ -1,7 +1,0 @@
-package service
-
-import "github.com/sourcegraph/sourcegraph/internal/database/dbtesting"
-
-func init() {
-	dbtesting.DBNameSuffix = "batchesservicedb"
-}