--- a/enterprise/internal/batches/store/main_test.go
+++ b/enterprise/internal/batches/store/main_test.go
@@ -1,7 +1,0 @@
-package store
-
-import "github.com/sourcegraph/sourcegraph/internal/database/dbtesting"
-
-func init() {
-	dbtesting.DBNameSuffix = "batchchangesstoredb"
-}