--- a/enterprise/internal/batches/webhooks/main_test.go
+++ b/enterprise/internal/batches/webhooks/main_test.go
@@ -11,14 +11,9 @@ import (
 	"testing"
 
 	btypes "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/types"
-	"github.com/sourcegraph/sourcegraph/internal/database/dbtesting"
 	"github.com/sourcegraph/sourcegraph/internal/repos"
 	"github.com/sourcegraph/sourcegraph/internal/types"
 )
-
-func init() {
-	dbtesting.DBNameSuffix = "batchchangesswebhooksdb"
-}
 
 var update = flag.Bool("update", false, "update testdata")